Updated Scripting Engines:
Groovy to 2.1.4:  All tests passed with no changes
Beanshell to: 2.0b5: All tests passed with no changes
Javascript Rhino to: 1.7R4  All tests passed with no changes
Jython:  2.7-b1.  All uses of Parameter.ZERO had to be changed to DefaultConstructorParameter.INSTANCE for the scripts to compile. (I'm assuming var-args bug in script engine)
Jruby to 1.7.4:  This caused the most problems.  To use traditional Pico composition with .addComponent(A), every class reference had to have .java_class appended to it for things to work.  Anybody know of a way to work around this for the latest JRuby??  (So its now .addComponent(A.java_class);